The start of a beautiful friendship
This code is not production ready.  Do not try to compile and connect this to the AIQ network, as shit is bound to break.  It will gobble your wallet, so do not use your existing wallet.dat file with this for the time being.  Still being worked on, and this is just to keep the punters satisfied.  Continual work will be done in the coming days/weeks.