Added fix to connect to AIQ Network
Artiqox legacy client featured a borked CoinbaseMaturity model that we need to carry forward with into the future as it is now part of the Blockchain.  A workaround has been added to the modern client to get it to connect and accept blocks.  Blocks from 145000 are what we are calling AuxPoW2 for the sake of.